Several small improvements/changes implemented during the Semantic Interaction Hackatron in Vienna
Changed the Component names for Dereferencers and Searchers from 
  "org.apache.stanbol.entityhub.site.<ImplClassName>"
to
  "org.apache.stanbol.entityhub.dereferencer.<ImplClassName>"
  "org.apache.stanbol.entityhub.searcher.<ImplClassName>"

This prepares a planed change to allow configuring the Dereferencer/Searcher for a ReferencedSIte only by using the simple Class name (instead of the full name). The initialisation of the Component within the ReferencedSite will than also look for "searcher.<ImplClassName>" and "dereferencer.<ImplClassName>" if no Component for the configured value can be instantiated by the available ComponentFactories.

Changed the name of the constant DEFAULT_MAPEED_ENTITY_STATE to DEFAULT_MAPPING_STATE

SolrYard: Changed all select requests from GET to POST after discovering some problems with the URL Encoding of special Characters like 'é' when using GET

Corrected a bug of the RdfIndexer  The indexer collects <chunksize> representations in a local cache to index them all at once. The Bug caused Representations remaining in the cache when the indexing completed to be not indexed in the SolrYard.

dblp indexer: Added the option that allows to configure the indexer to use MultiYardLayout (see SolrYard documentation for details)

DBpedia indexer: Added a readme.txt that describes the basic steps needed
to use this indexer.

Removed an unused metatype.properties form the linkedData bundle

git-svn-id: https://svn.apache.org/repos/asf/incubator/stanbol/trunk@1074848 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: ade106f96dd6acf9a485edaa260913da405725d7