STANBOL-141: The last missing things to support for Installing SolrIndexes based on Archives.
Added support for Index Data References (Index Archives that just contain a link to the actual data). The reference to the index data is resolved by the DataFileProvider Service. The SolrYard now keeps track of "uninitialised indexes" - this are indexes that where installed, but could not be initialised because the required data where not available via the DataFileProvider service. 

 - Updated the SolrDirectoryManager interface to deal with "uninitialized" indexes.
 - Added a parameter to the SolrYardConfig that allows to specify if the Solr Index is allowed to be initialised with the default schema (and without data).
 - The DefaultSolrDirectoryManager now keeps a persistent list of "uninitialised" indexes. This means that index data references (property files) are stored within the data directory of the SolrYard bundle so that they can be resent to the DataFileProvider service when the initialisation of the index is requested the next time.

This functionality requires an OSGI environment. In case the SolrYard is running outside an OSGI Environment such methods will throw an IllegalStateException with an according message.
 
Note that directly including the index data in the solr archive is still supported and still the preferred way to go for small or empty indexes. Empty indexes can be used to use a special Solr Schema/Configuration for the initialisation of an yard.



git-svn-id: https://svn.apache.org/repos/asf/incubator/stanbol/trunk@1091050 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 5dc6a69ccf22b0b8f6c232bfc8b442a1274a63a6