STANBOL 306:
-Added an implementation which goes through annotated RDF data and updates CMIS repository accordingly. 

Unlike the JCR, this implementation is not able to create an hierarchy from documents. Instead, it tries to create folder hierarchy from the path of top root documents and all children documents are created within the same folder with the root.
 
Custom annotations retrieved from RDF data are obtained for each document as stored as
ContentStream of document.

git-svn-id: https://svn.apache.org/repos/asf/incubator/stanbol/trunk@1154755 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 17966cbbcf32d8e69cb2be1e862d4477d905d4d8