STANBOL-105: Added Support for CORS Allow-Origin and preflight requests.
The Origin globally allowed by Stanbol are configured with the JesreyEndpoint component. If '*' is configured, than all origins are allowed and "Access-Control-Allow-Origin: * " will be returned in CORS enabled responses. Otherwise the Origin is checked with equals to the configured list.

JerseyResources need to explicitly support CORS. To do that the static helper methods of the new CorsHelper utility class should be used.


other changes:

* added metatype.properties for the JeryeyEndpoint component
* removed an unused import to scala.actors.threadpool.Arrays
* added Concepts and Others to the OR arguments checking if the Extracted Entities header should be added in the ContentItem Freemaker template

git-svn-id: https://svn.apache.org/repos/asf/incubator/stanbol/trunk@1178354 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: d8206002ab9d3b9f6e71ae4a7f14a819de38641d