STANBOL-364: Missing files to commit
Added implementation for the CMS Adapter feature that provides submission of content repository objects to Contenthub. The main interface offering the functionality is ContenthubFeeder. There are two implementations of it for JCR and CMIS content repositories.

These implementations basically create SolrContentItems from JCR nodes or CMIS documents and submit them to the Contenthub. Properties of these content repository objects are also indexed in Solr as external metadata together with the actual content. Indexed properties are used to provide faceted search over the submitted documents.

It is also possible to delete content items from Contenthub based on the content repository IDs or paths of nodes/documents.

git-svn-id: https://svn.apache.org/repos/asf/incubator/stanbol/trunk@1189810 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 724e094fad524b143e11c1d0f85fd7d6dea2719e