Bug fix:
This fixes a thread synchronisation Issue during the initialisation phase (line 145-153 EnhancementJobHandler).
The first call to "executeNextNodes(..)" - async. calls EnhancementEngines that are not dependendOn any other - was not executed within a write lock.

Therefore in cases multiple EnhancementEngines could be started there was the possibility that the first one completed before the second one was correctly initialised. In such cases it could happen, that the JobManager attempt to start the same engine twice. This caused an IllegalStateException.

Now the  "executeNextNodes(..)" during the initialisation is correctly issued within a write lock.

Other changes:

* improved debug level logging messages to be more informative

git-svn-id: https://svn.apache.org/repos/asf/incubator/stanbol/trunk@1294670 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 3bc81f04bd5a39321ab926d6697372a715f1bb72