no STANBOL issue, because Jira is down since some time.
Removed the dependency of the Enhancer Web interface to the TcManager. For the SPARQL endpoint to be available there is still the Clerezza QueryEngine required. However this dependency is no OPTIONAL. If no QueryEngine Implementation is available SPARQL requests will return "503 - Service unavailable" with an according error message.

Also organized imports of all classes

git-svn-id: https://svn.apache.org/repos/asf/incubator/stanbol/trunk@1356754 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: e2740f1de5be850998a18f1e42d796546929c5a9