Trying to commit the big patch prepared by Meric, in smaller chunks in accordance to the opened issues. In this commit, I am trying to commit changes to adapt multipart serialization of ContentItems. So, it mainly includes fixes for STANBOL-785, STANBOL-786 and STANBOL-788.
As explained in the mentioned issues, adapting the multipart serialization of ContentItems, we do not need constraint parameter to be passed explicitly. Instead additional metadata comes as a certain part of the ContentItem. Based on this, the "constraint" parameters are removed from the RESTful services of StoreResource.

Based on these changes, documentation of the services and the HTML interface are also updated. 

There is also an additional RESTful service, taking a ContentItem parameter which was created by the ContentItemReader (i.e providing the multipart deserialization of ContentItems)

git-svn-id: https://svn.apache.org/repos/asf/stanbol/trunk@1414032 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 8d9253b831a42ccf7817f20f844add635ce8efa6