Fix for STANBOL-782.
It is not mandatory for a ContentItem to have a textual content part i.e a content part having one of the TEXT_HTML, TEXT_PLAIN or TEXT_XML media types. If it has a binary content only the binary content of the ContentItem is stored but not indexed. A StoreException is thrown only if the ContentItem does have neither a textual content part or any binary content.

git-svn-id: https://svn.apache.org/repos/asf/stanbol/trunk@1421355 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 405594d410d90990299a7149504b357a3aafce20